added subject property detail.

+++ Rainmaker2/src/MobileApps/Android/Colaba-Android/app/src/main/java/com/rnsoft/colabademo/activities/model/SubjectPropertyDetails.kt
@@ -18,7 +18,7 @@ data class SubPropertyData(
     var loanApplicationId : Int,
     var propertyTypeId : Int,
     var occupancyTypeId : Int,
-    var appraisedPropertyValue : Float?,
+    var appraisedPropertyValue : Double?,
     var propertyTax : Double?,
     var homeOwnerInsurance : Double?,
     var floodInsurance : Double?,
